reproduce and fix overflow error in TxAddInput
Also add a `>=0` requirement on `varint` codec input.

+++ src/commonMain/kotlin/fr/acinq/lightning/serialization/v4/Deserialization.kt
@@ -177,7 +177,7 @@ object Deserialization {
         serialId = readNumber(),
         outPoint = readOutPoint(),
         txOut = TxOut.read(readDelimitedByteArray()),
-        sequence = readNumber()
+        sequence = readNumber().toUInt()
     )
 
     private fun Input.readRemoteTxAddOutput() = RemoteTxAddOutput(
